+++ src/main/php/lang/reflection/Constructor.class.php
@@ -13,13 +13,7 @@ class Constructor extends Routine {
 
   /** @return string */
   public function toString() {
-
-    // Parse apidoc. FIXME!
-    preg_match_all('/@(param)\s+(.+)/', $this->reflect->getDocComment(), $matches, PREG_SET_ORDER);
-    $tags= [];
-    foreach ($matches as $match) {
-      $tags[$match[1]][]= $match[2];
-    }
+    $tags= Reflection::meta()->tags($this->reflect);
 
     // Compile signature
     $sig= '';