+++ src/main/php/lang/reflection/Method.class.php
@@ -42,25 +42,25 @@ class Method extends Routine {
     }
   }
 
-  /** @return lang.reflection.TypeHint */
+  /** @return lang.reflection.Constraint */
   public function returns() {
     $t= $this->reflect->getReturnType();
     if (null === $t) {
       $present= false;
 
-      // Check for type in api documentation, defaulting to `var`
+      // Check for type in meta information, defaulting to `var`
       $t= Type::$VAR;
     } else if ($t instanceof \ReflectionUnionType) {
       $union= [];
       foreach ($t->getTypes() as $component) {
         $union[]= Type::resolve($component->getName(), $this->resolver());
       }
-      return new TypeHint(new TypeUnion($union));
+      return new Constraint(new TypeUnion($union));
     } else {
       $name= PHP_VERSION_ID >= 70100 ? $t->getName() : $t->__toString();
 
       // Check array, self and callable for more specific types, e.g. `string[]`,
-      // `static` or `function(): string` in api documentation
+      // `static` or `function(): string` in meta information
       if ('array' === $name) {
         $t= Type::$ARRAY;
       } else if ('callable' === $name) {
@@ -68,33 +68,22 @@ class Method extends Routine {
       } else if ('self' === $name) {
         $t= new XPClass($this->reflect->getDeclaringClass());
       } else {
-        return new TypeHint(Type::resolve($name, $this->resolver()));
+        return new Constraint(Type::resolve($name, $this->resolver()));
       }
       $present= true;
     }
 
-    // Parse apidoc. FIXME!
-    preg_match_all('/@(return|param)\s+(.+)/', $this->reflect->getDocComment(), $matches, PREG_SET_ORDER);
-    $tags= [];
-    foreach ($matches as $match) {
-      $tags[$match[1]][]= rtrim($match[2], ' */');
-    }
-
-    return new TypeHint(
-      isset($tags['return'])? Type::resolve($tags['return'][0], $this->resolver()) : $t,
+    // Use meta information
+    $this->meta ?? $this->meta= $this->meta();
+    return new Constraint(
+      isset($this->meta[DETAIL_RETURNS]) ? Type::resolve($this->meta[DETAIL_RETURNS], $this->resolver()) : $t,
       $present
     );
   }
 
   /** @return string */
   public function toString() {
-
-    // Parse apidoc. FIXME!
-    preg_match_all('/@(return|param)\s+(.+)/', $this->reflect->getDocComment(), $matches, PREG_SET_ORDER);
-    $tags= [];
-    foreach ($matches as $match) {
-      $tags[$match[1]][]= $match[2];
-    }
+    $tags= Reflection::meta()->tags($this->reflect);
 
     // Compile signature
     $sig= '';